+++ lib/vcs_git_blocking/test/test__path_resolution.mli
@@ -18,13 +18,3 @@
 (*_  and the LGPL-3.0 Linking Exception along with this library. If not, see    *)
 (*_  <http://www.gnu.org/licenses/> and <https://spdx.org>, respectively.       *)
 (*_******************************************************************************)
-
-type t
-
-include Vcs_git_provider.Runtime.S with type t := t
-
-val create : unit -> t
-
-module Private : sig
-  val find_executable : path:string -> string option
-end