+++ CHANGES.md
@@ -1,3 +1,19 @@
+## 0.0.12 (unreleased)
+
+### Added
+
+### Changed
+
+- Pre-locate the git executable in `vcs_git_blocking` (#52, @mbarbin).
+
+### Deprecated
+
+### Fixed
+
+- Dispose of `Shexp_process.Context` in `vcs_git_blocking` (#52, @mbarbin).
+
+### Removed
+
 ## 0.0.11 (2025-04-13)
 
 ### Changed